feat: Display the information of course student in another window. 80% of student activities have been finished

+++ SFML/Course.h
@@ -50,7 +50,7 @@ public:
         this->setMaxStu(maxStu);
 
         scoreFile.readFile();
-        int stuNum = scoreFile.cnt.size();
+        long long stuNum = scoreFile.cnt.size();
         if (stuNum > this->getMaxStu()) {
             displayErrorExceedMaxStu();
         }
